feat: 저장한 동영상이 없을 때 메인 페이지 구현
1. 뷰를 세분화
  - 기존의 뷰는 modal에 관련한 부분에 대부분으로 modalView로 별도 분리
  - 추가된 뷰는 저장한 영상 관련으로 SavedVideosView로 분리
2. 저장 없음 페이지 구현
  - 페이지 렌더링 시 localStorage에 항목이 없을 때 보여주는 부분 구현
  - 검색 후 저장 한 뒤 모달을 닫으면 자동으로 업데이트해서 위에서 표시한 항목
    을 제거하도록 구현
