+++ index.html
@@ -23,7 +23,7 @@
       <section class="saved-videos">
         <h3 hidden>저장한 비디오 목록</h3>
         <ul class="video-list">
-          <li class="video-item">
+          <!-- <li class="video-item">
             <img
               src="https://i.ytimg.com/vi/mQhgF7RoUCA/mqdefault.jpg"
               alt="video-item-thumbnail"
@@ -49,7 +49,7 @@
                 🗑
               </button>
             </div>
-          </li>
+          </li> -->
         </ul>
       </section>
     </main>
